+++ tests/helpers.py
@@ -41,6 +41,13 @@ TEST_POINTCLOUD_URLS = [
 
 TEST_VIDEO_URL = "https://github.com/scaleapi/nucleus-python-client/raw/master/tests/testdata/video.mp4"
 
+TEST_INACCESSIBLE_VIDEO_URL = "https://github.com/scaleapi/nucleus-python-client/raw/master/tests/testdata/video.mp4/fake-for-privacy-mode"
+
+TEST_INACCESSIBLE_IMG_URLS = [
+    "https://github.com/scaleapi/nucleus-python-client/raw/master/tests/testdata/airplane.jpeg/fake-for-privacy-mode",
+    "https://github.com/scaleapi/nucleus-python-client/raw/master/tests/testdata/arctichare.jpeg/fake-for-privacy-mode",
+]
+
 TEST_LIDAR_SCENES = {
     "scenes": [
         {
@@ -90,7 +97,6 @@ TEST_VIDEO_SCENES = {
     "scenes": [
         {
             "reference_id": "scene_1",
-            "video_upload_type": "image",
             "frame_rate": 15,
             "frames": [
                 {
@@ -110,9 +116,29 @@ TEST_VIDEO_SCENES = {
         },
         {
             "reference_id": "scene_2",
-            "video_upload_type": "video",
             "video_url": TEST_VIDEO_URL,
             "metadata": {"timestamp": "1234", "weather": "rainy"},
+        },
+        {
+            "reference_id": "scene_3",
+            "video_url": TEST_INACCESSIBLE_VIDEO_URL,
+            "frame_rate": 15,
+            "frames": [
+                {
+                    "image_url": TEST_INACCESSIBLE_IMG_URLS[0],
+                    "type": "image",
+                    "reference_id": "video_frame_2",
+                    "metadata": {"time": 123, "foo": "bar"},
+                },
+                {
+                    "image_url": TEST_INACCESSIBLE_IMG_URLS[1],
+                    "type": "image",
+                    "reference_id": "video_frame_3",
+                    "metadata": {"time": 124, "foo": "bar_2"},
+                },
+            ],
+            "metadata": {"timestamp": "1234", "weather": "rainy"},
+            "upload_to_scale": False,
         },
     ],
     "update": False,
@@ -122,7 +148,6 @@ TEST_VIDEO_SCENES_REPEAT_REF_IDS = {
     "scenes": [
         {
             "reference_id": "scene_1",
-            "video_upload_type": "image",
             "frame_rate": 15,
             "frames": [
                 {
@@ -148,7 +173,6 @@ TEST_VIDEO_SCENES_INVALID_URLS = {
     "scenes": [
         {
             "reference_id": "scene_1",
-            "video_upload_type": "image",
             "frame_rate": 15,
             "frames": [
                 {
@@ -168,13 +192,11 @@ TEST_VIDEO_SCENES_INVALID_URLS = {
         },
         {
             "reference_id": "scene_2",
-            "video_upload_type": "video",
             "video_url": TEST_IMG_URLS[0],
             "metadata": {"timestamp": "1234", "weather": "rainy"},
         },
         {
             "reference_id": "scene_3",
-            "video_upload_type": "video",
             "video_url": TEST_VIDEO_URL + "nonsense",
             "metadata": {"timestamp": "1234", "weather": "rainy"},
         },